+++ modules/autofill_object.py
@@ -1,1 +1,2 @@
+from modules.autofill_object_popup import *
 from modules.autofill_object_save_info import *
