+++ modules/data/address_fill.components.json
@@ -1,0 +1,13 @@
+{
+    "form-field": {
+        "selectorData": "input[autocomplete=\"{name}\"]",
+        "strategy": "css",
+        "groups": []
+    },
+
+    "submit-button": {
+        "selectorData": "input[type=\"{name}\"]",
+        "strategy": "css",
+        "groups": []
+    }
+}
