+++ tests/form_autofill/test_enable_disable_autofill.py
@@ -1,9 +1,11 @@
 import pytest
 from selenium.webdriver import Firefox
 
-from modules.autofill_object import AutofillSaveInfo
+from modules.autofill_object import AddressFill
+from modules.autofill_object_popup import AutofillPopup
 from modules.browser_object import Navigation
 from modules.page_object import AboutPrefs
+from modules.util import Utilities
 
 countries = ["CA", "US"]
 
@@ -14,11 +16,23 @@ def test_enable_disable_autofill(driver: Firefox, country_code: str):
     C122347, tests that after filling autofill and disabling it in settings that
     the autofill popups do not appear.
     """
+    # instantiate objects
     Navigation(driver).open()
-    afsi = AutofillSaveInfo(driver).open()
-    autofill_sample_data_canadian = afsi.fake_autofill_data(country_code)
-    afsi.save_information_basic(autofill_sample_data_canadian)
+    af = AddressFill(driver).open()
+    afp = AutofillPopup(driver)
+    util = Utilities()
+
+    # create fake data, fill it in and press submit and save on the doorhanger
+    autofill_sample_data = util.fake_autofill_data(country_code)
+    af.save_information_basic(autofill_sample_data)
+    afp.press_doorhanger_save()
     about_prefs = AboutPrefs(driver, category="privacy").open()
-    about_prefs.find_setting_and_click("save-and-fill-addresses")
-    new_afsi = AutofillSaveInfo(driver).open()
-    new_afsi.double_click_name_and_verify()
+    about_prefs.get_element("save-and-fill-addresses").click()
+
+    # creating new objects to prevent stale webelements
+    new_af = AddressFill(driver).open()
+    new_afp = AutofillPopup(driver)
+
+    # verifying the popup panel does not appear
+    new_af.double_click("form-field", "name")
+    new_afp.verify_no_popup_panel()
